Terminado operaciones:
Añadidos test
Modificados los operandos por enteros
Corregido error en multiplication

+++ src/main/java/es/operaciones/CarlosAzanon/Operation.java
@@ -2,23 +2,23 @@ package es.operaciones.CarlosAzanon;
 
 public abstract class Operation {
     
-    private double operador1;
-    private double operador2;
+    private int operador1;
+    private int operador2;
     
-    public Operation(double op1,double op2){
+    public Operation(int op1,int op2){
         operador1=op1;
         operador2=op2;
     }
     
-    public abstract double calcular();
+    public abstract int calcular();
     
     public abstract String toString();
     
-    public double getOperador1(){
+    public int getOperador1(){
         return operador1;
     }
     
-    public double getOperador2(){
+    public int getOperador2(){
         return operador2;
     }
